+++ src/app/admin/components/surveillance/editSurveillance.controller.spec.js
@@ -75,6 +75,11 @@
                 expect(vm.missingEndDate()).toBe(true);
             });
 
+            it('should require an end date when there are no NCs and there\'s no surveillance end date', function () {
+                vm.surveillance.requirements[0].nonconformities = [];
+                expect(vm.missingEndDate()).toBe(true);
+            });
+
             it('should not require an end date when all NCs are closed and the surveillance has an end date', function () {
                 vm.surveillance.requirements[0].nonconformities[0].status = {id:2, name: 'Closed'};
                 vm.surveillance.endDateObject = '1472702800000';